Fix identity tests.

+++ KittenWeather/Temperature.m
@@ -8,15 +8,15 @@
 
 #import "Temperature.h"
 
-const float KELVIN_TO_CELSIUS_OFFSET = 273.15;
-const float KELVIN_TO_IMPERIAL_OFFSET = 459.67;
+const long double KELVIN_TO_CELSIUS_OFFSET = 273.15;
+const long double KELVIN_TO_IMPERIAL_OFFSET = 459.67;
 
-const float KELVIN_TO_IMPERIAL_RATIO = 9.0 / 5.0;
-const float IMPERIAL_TO_KELVIN_RATIO = 5.0 / 9.0;
+const long double KELVIN_TO_IMPERIAL_RATIO = 9.0 / 5.0;
+const long double IMPERIAL_TO_KELVIN_RATIO = 5.0 / 9.0;
 
 @interface Temperature()
 
-@property (nonatomic, assign) float privateKelvinDegrees;
+@property (nonatomic, assign) long double privateKelvinDegrees;
 
 @end
